feat: back-end with more functions

+++ my-backend/controllers/inventory.controller.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const Inventory = require('../models/inventory.model');
 
 exports.getAllItems = async (req, res) => {
@@ -11,8 +12,13 @@ exports.getAllItems = async (req, res) => {
 
 exports.getItemById = async (req, res) => {
   try {
+    // Verifica se o ID é um ObjectId válido
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ message: 'ID inválido' });
+    }
+
     const item = await Inventory.findById(req.params.id);
-    if (!item) return res.status(404).json({ message: 'Item not found' });
+    if (!item) return res.status(404).json({ message: 'Item não encontrado' });
     res.json(item);
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -31,8 +37,13 @@ exports.createItem = async (req, res) => {
 
 exports.updateItem = async (req, res) => {
   try {
+    // Verifica se o ID é um ObjectId válido
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ message: 'ID inválido' });
+    }
+
     const updatedItem = await Inventory.findByIdAndUpdate(req.params.id, req.body, { new: true });
-    if (!updatedItem) return res.status(404).json({ message: 'Item not found' });
+    if (!updatedItem) return res.status(404).json({ message: 'Item não encontrado' });
     res.json(updatedItem);
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -41,8 +52,13 @@ exports.updateItem = async (req, res) => {
 
 exports.deleteItem = async (req, res) => {
   try {
+    // Verifica se o ID é um ObjectId válido
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ message: 'ID inválido' });
+    }
+
     const deletedItem = await Inventory.findByIdAndDelete(req.params.id);
-    if (!deletedItem) return res.status(404).json({ message: 'Item not found' });
+    if (!deletedItem) return res.status(404).json({ message: 'Item não encontrado' });
     res.status(204).send();
   } catch (err) {
     res.status(500).json({ message: err.message });
